Fix compiler warnings
* conversion-null: converting to non-pointer type from NULL
* int-to-pointer-cast: cast to pointer from integer of different size
* pointer-arith: NULL used in arithmetic

+++ runtime/compiler/x/codegen/CallSnippet.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2018 IBM Corp. and others
+ * Copyright (c) 2000, 2019 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -953,7 +953,7 @@ uint8_t *TR::X86CallSnippet::emitSnippetBody()
       relocation = new (cg()->trHeapMemory()) TR::ExternalRelocation(
             cursor,
             *(uint8_t **)cursor,
-             getNode() ? (uint8_t *)getNode()->getInlinedSiteIndex() : (uint8_t *)-1,
+             getNode() ? (uint8_t *)(uintptr_t)getNode()->getInlinedSiteIndex() : (uint8_t *)-1,
             TR_ConstantPool,
             cg());
 
@@ -1137,7 +1137,7 @@ uint8_t *TR::X86UnresolvedVirtualCallSnippet::emitSnippetBody()
    cg()->addExternalRelocation(
       new (cg()->trHeapMemory()) TR::ExternalRelocation(cursor,
                                                             (uint8_t *)cpAddr,
-                                                            getNode() ? (uint8_t *)getNode()->getInlinedSiteIndex() : (uint8_t *)-1,
+                                                            getNode() ? (uint8_t *)(uintptr_t)getNode()->getInlinedSiteIndex() : (uint8_t *)-1,
                                                             TR_ConstantPool,
                                                             cg()),
                  __FILE__, __LINE__, getNode());
